Move the clean up logic to platform specific initializers
