+++ lib/init-electron.js
@@ -7,6 +7,7 @@ const electron = require('./electron/electron')
 
 const paths = require('./paths')
 const UserDefaults = require('./UserDefaults')
+const WorkspaceRegistry = require('./WorkspaceRegistry')
 
 const Logger = require('./Logger')
 const logger = new Logger({ name: 'init-electron' })
@@ -37,4 +38,15 @@ function writeUserDeafults () {
 
 electron.app.once('will-quit', () => {
   writeUserDeafults()
+})
+
+/*
+Setup listeners for new workspaces
+in order to remove any dangling
+references
+*/
+WorkspaceRegistry.getInstance().on('add', async workspace => {
+  workspace.on('cleanup', async () => {
+    workspace.cleanupSockets()
+  })
 })
